1605: Refactor View case page
