+++ accessibility_monitoring_platform/apps/common/templates/common/view_accordion_section.html
@@ -1,4 +1,4 @@
-<div class="govuk-accordion__section">
+<div class="govuk-accordion__section" data-search-target-page-name="{{ view_section.name }}" data-search-target-label="Edit" data-search-target-url="{{ view_section.edit_url }}">
     <div class="govuk-accordion__section-header">
         <h2 id="{{ view_section.anchor }}" class="govuk-accordion__section-heading">
             <span class="govuk-accordion__section-button" id="accordion-heading-{{ view_section.name|slugify }}">
